refactor: 한글 필터 로직 모델에서 분리, extension String 으로 변경
